4. GREEN - pass by properly casting to double before doing divisions and by rounding back to int on completion.

+++ Day13/AoC-Day13/src/day13.h
@@ -1,11 +1,11 @@
 typedef struct Game {   
-  int deltaXA;
-  int deltaXB;
-  int deltaYA;
-  int deltaYB;
-  int targetX;
-  int targetY;        
+  long deltaXA;
+  long deltaXB;
+  long deltaYA;
+  long deltaYB;
+  long targetX;
+  long targetY;        
 } Game;
 
-int cost_to_win_game(Game game);
-int cost_to_win_from_instructions(char* instructions);
+long cost_to_win_game(Game game);
+long cost_to_win_from_instructions(char* instructions);
